Добавлена логика комиссии и обновлена статистика

+++ backend/techwizBackend/pkg/models/User.go
@@ -17,8 +17,8 @@ type User struct {
 	Status      string          `json:"status,omitempty" bson:"status,omitempty"`         // only master // default, senior or premium
 	Category    []Category      `json:"categories,omitempty" bson:"categories,omitempty"` // only master
 	CategoryId  []bson.ObjectID `json:"categories_id,omitempty" bson:"categories_id,omitempty"`
-	Balance     float32         `json:"balance,omitempty" bson:"balance,omitempty"`       // only master
-	Commission  int             `json:"commission,omitempty" bson:"commission,omitempty"` // only master
+	Balance     float64         `json:"balance,omitempty" bson:"balance,omitempty"`       // only master
+	Commission  float64         `json:"commission,omitempty" bson:"commission,omitempty"` // only master
 	Dismissed   bool            `json:"dismissed" bson:"dismissed"`
 }
